requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
cryptography==43.0.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt
openpyxl==3.1.5
python-multipart==0.0.12
alembic==1.14.0
pandas==2.2.3
xlrd==2.0.1
azure-storage-blob==12.23.1
